requirements-kaggle.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0
Pillow>=9.5.0
timm>=0.9.0
diffusers>=0.21.0
accelerate>=0.24.0
torchdiffeq>=0.2.3
matplotlib>=3.7.0
scikit-learn>=1.3.0
seaborn>=0.13.0
umap-learn>=0.5.6
PyWavelets>=1.4.1
